Sig-instrumentation e2e tests refactoring
