+++ test/e2e/BUILD
@@ -21,8 +21,9 @@ go_test(
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//test/e2e/apimachinery:go_default_library",
         "//test/e2e/autoscaling:go_default_library",
-        "//test/e2e/cluster-logging:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/instrumentation/logging:go_default_library",
+        "//test/e2e/instrumentation/monitoring:go_default_library",
         "//test/e2e/kubectl:go_default_library",
         "//test/e2e/metrics:go_default_library",
         "//test/e2e/scalability:go_default_library",
@@ -48,7 +49,6 @@ go_library(
         "addon_update.go",
         "apparmor.go",
         "audit.go",
-        "cadvisor.go",
         "certificates.go",
         "cluster_upgrade.go",
         "dashboard.go",
@@ -65,13 +65,10 @@ go_library(
         "gke_node_pools.go",
         "ha_master.go",
         "ingress.go",
-        "kibana_logging.go",
         "kube_proxy.go",
         "kubelet.go",
         "kubelet_perf.go",
         "limit_range.go",
-        "logging_soak.go",
-        "monitoring.go",
         "network_partition.go",
         "network_policy.go",
         "networking.go",
@@ -93,7 +90,6 @@ go_library(
         "service_latency.go",
         "serviceloadbalancers.go",
         "ssh.go",
-        "stackdriver_monitoring.go",
         "ubernetes_lite.go",
         "util_iperf.go",
     ],
@@ -134,14 +130,11 @@ go_library(
         "//test/utils/junit:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/google/cadvisor/info/v1:go_default_library",
-        "//vendor/github.com/influxdata/influxdb/client/v2:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/ginkgo/config:go_default_library",
         "//vendor/github.com/onsi/ginkgo/reporters:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/golang.org/x/crypto/ssh:go_default_library",
-        "//vendor/golang.org/x/oauth2/google:go_default_library",
-        "//vendor/google.golang.org/api/monitoring/v3:go_default_library",
         "//vendor/k8s.io/api/batch/v1:go_default_library",
         "//vendor/k8s.io/api/batch/v2alpha1:go_default_library",
         "//vendor/k8s.io/api/certificates/v1beta1:go_default_library",
@@ -195,10 +188,10 @@ filegroup(
         "//test/e2e/apimachinery:all-srcs",
         "//test/e2e/autoscaling:all-srcs",
         "//test/e2e/chaosmonkey:all-srcs",
-        "//test/e2e/cluster-logging:all-srcs",
         "//test/e2e/common:all-srcs",
         "//test/e2e/framework:all-srcs",
         "//test/e2e/generated:all-srcs",
+        "//test/e2e/instrumentation:all-srcs",
         "//test/e2e/kubectl:all-srcs",
         "//test/e2e/manifest:all-srcs",
         "//test/e2e/metrics:all-srcs",
