+++ test/e2e/instrumentation/OWNERS
@@ -1,0 +1,9 @@
+approvers:
+- crassirostris
+- DirectXMan12
+- fabxc
+- fgrzadkowski
+- piosz
+- x13n
+reviewers:
+- sig-instrumentation-pr-reviews
