+++ test/e2e/instrumentation/logging/sd_events.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package logging
 
 import (
 	"time"
@@ -22,6 +22,7 @@ import (
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/instrumentation"
 
 	. "github.com/onsi/ginkgo"
 )
@@ -40,7 +41,7 @@ const (
 	eventCreationInterval = 10 * time.Second
 )
 
-var _ = framework.KubeDescribe("Cluster level logging implemented by Stackdriver", func() {
+var _ = instrumentation.SIGDescribe("Cluster level logging implemented by Stackdriver", func() {
 	f := framework.NewDefaultFramework("sd-logging-events")
 
 	BeforeEach(func() {
